requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
pycord==2.0.0b5
idna==2.8
multidict==5.1.0
PyYAML==5.3.1
typing-extensions==3.10.0.0
yarl==1.6.3